componetizando el youtube components y pasando por props las propiedades los tests en forma de array y el ID del video como para reutilizarlo cada vez que se requiera
